refactor: 좋아요 삭제 service 존재하지 않는 좋아요를 요청할때 예외처리 추가
- postId와 userId로 likeId를 찾을때 존재 하지 않으면 null을 반환
- 만약 null이면 클라이언트에서 잘못된 값으로 요청한것이므로 INVALID_PARAMETER_VALUE 예외 발생

+++ backend/src/main/java/our/portfolio/devspace/domain/like/service/LikeService.java
@@ -34,6 +34,7 @@ public class LikeService {
     public CreateLikeResponse createLike(CreateLikeRequest likeRequest, Long userId) {
 
         Like like = likeMapper.toEntity(likeRequest, userId);
+
         likeRepository.save(like);
 
         return likeMapper.toCreateLikeResponse(like);
@@ -41,6 +42,10 @@ public class LikeService {
 
     public DeleteLikeResponse deleteLike(Long postId, Long userId) {
         Long likeId = likeRepository.findIdByProfileIdAndPostId(postId, userId);
+
+        if (likeId == null) {
+            throw new CustomException(INVALID_PARAMETER_VALUE);
+        }
         likeRepository.deleteById(likeId);
 
         return likeMapper.toDeleteLikeResponse(postId);
